Move the imports of external inventories to its own class
Code regarding the Carma inventory is moved to a separate class.
DatabaseCleaner and NewDatabase classes are modified accordingly.
